feat: Menu 객체에서 InputView 분리
