+++ src/main/java/controller/state/GameState.java
@@ -1,11 +1,9 @@
 package controller.state;
 
-import domain.Square;
-
 public interface GameState {
     GameState start();
 
-    GameState play(Square source, Square target);
+    GameState play();
 
     GameState status();
